refactor(contracts): extract task result types
